pyproject.toml
[project]
name = "persona"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=3.1.0",
    "openai>=1.57.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "tqdm>=4.67.1",
    "transformers>=4.47.0",
]
